quiet spec "does not exist" and "already exists" messages
The development and test db names were the same, and rake tries to
create both by default if they're local, so the first try would
succeed and the second would fail.  Didn't seem to break the tests but
output lots of messages if logging is enabled.

Changed the development and test db names so they're different.  Now
both can be created and dropped so rake is quiet.