fix Rails4/Bundler binary stubs conflict
Rails 4 doesn't like the stubs that Bundler generates when you run
"bundle binstubs" and it crashes with a message like:

 Looks like your app's ./bin/rails is a stub that was generated by Bundler.

It appears that we weren't using the stubs anyway, so I removed the
binstubs invocation, but if we need them the new syntax to generate
Rails4-compatible stubs is "bundle binstubs rails".  You can add other
stubs after rails if you need them.