query the adapters' default pk instead of assuming its value
The mysql adapter's default primary key definition changed in 4.1 so we
can't assume that it's a fixed string anymore.  We now query the adapter
before we start testing and use that value to check that we haven't
modified anything.

NOTE: we'll need to be more subtle about changing the default PK value,
too.  We should modify the PK string instead of replacing it with a
hard-coded one.